ai-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and Async
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Image Processing
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
scikit-image==0.22.0

# AI/ML Libraries
openai==1.6.1
replicate==0.22.0
rembg==2.0.52  # Local background removal
transformers==4.36.2  # For local AI models
torch==2.1.2  # PyTorch for AI models
torchvision==0.16.2

# Translation
deepl==1.16.1
googletrans==4.0.0-rc1

# Utilities
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
python-jose[cryptography]==3.3.0

# Monitoring and Logging
sentry-sdk==1.39.1
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional GPU support
# For CUDA 11.8 (uncomment if using GPU)
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
# xformers==0.0.23  # Memory efficient transformers